# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Monitoring System API
  description: |
    API для системы мониторинга доступности REST endpoint'ов.

    MVP-функциональность:
    - регистрация сервисов с ответственными сотрудниками;
    - добавление endpoint'ов для мониторинга;
    - периодическая проверка endpoint'ов выполняется системой автоматически;
    - сохранение сырых результатов проверок;
    - отправка email-уведомлений при сбое;
    - получение истории проверок;
    - расчёт uptime и SLA за месяц;
    - получение сводной статистики по сервисам.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: System
    description: Служебные endpoint'ы Monitoring Service
  - name: Services
    description: Управление мониторируемыми сервисами
  - name: Endpoints
    description: Управление endpoint'ами сервисов
  - name: Checks
    description: История результатов проверок
  - name: Statistics
    description: Uptime, SLA и сводная статистика

paths:
  /health:
    get:
      tags:
        - System
      summary: Проверить работоспособность Monitoring Service
      description: Возвращает статус самого сервиса мониторинга.
      operationId: getHealth
      responses:
        "200":
          description: Monitoring Service доступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/services:
    post:
      tags:
        - Services
      summary: Зарегистрировать сервис с ответственными
      description: |
        Создаёт новый мониторируемый сервис.
        В рамках MVP ответственные сотрудники указываются сразу при создании сервиса.
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreateRequest"
            examples:
              createService:
                summary: Создание сервиса
                value:
                  name: orders-service
                  description: Сервис обработки заказов
                  slaTargetPercent: 99.9
                  responsibles:
                    - fullName: Иван Петров
                      email: ivan.petrov@example.com
                    - fullName: Анна Смирнова
                      email: anna.smirnova@example.com
      responses:
        "201":
          description: Сервис успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Services
      summary: Получить список сервисов
      description: Возвращает список зарегистрированных сервисов.
      operationId: getServices
      parameters:
        - name: isActive
          in: query
          required: false
          description: Фильтр по признаку активности сервиса
          schema:
            type: boolean
      responses:
        "200":
          description: Список сервисов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"

  /api/v1/services/{serviceId}:
    get:
      tags:
        - Services
      summary: Получить карточку сервиса
      description: Возвращает данные конкретного сервиса, включая ответственных сотрудников.
      operationId: getServiceById
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: Данные сервиса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Services
      summary: Обновить сервис
      description: |
        Обновляет параметры сервиса.
        Может использоваться для изменения описания, SLA или признака активности.
      operationId: updateService
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdateRequest"
      responses:
        "200":
          description: Сервис обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/services/{serviceId}/endpoints:
    post:
      tags:
        - Endpoints
      summary: Добавить endpoint сервиса
      description: |
        Добавляет REST endpoint, который будет проверяться системой мониторинга.
        В MVP endpoint проверяется методом GET.
      operationId: createEndpoint
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndpointCreateRequest"
            examples:
              createEndpoint:
                summary: Добавление endpoint'а
                value:
                  name: Orders health
                  url: http://test-service:8081/health/up
                  method: GET
                  timeoutMs: 2000
                  checkIntervalSec: 10
      responses:
        "201":
          description: Endpoint успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointResponse"
        "400":
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Endpoints
      summary: Получить endpoint'ы сервиса
      description: Возвращает список endpoint'ов, зарегистрированных для сервиса.
      operationId: getServiceEndpoints
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - name: isActive
          in: query
          required: false
          description: Фильтр по признаку активности endpoint'а
          schema:
            type: boolean
      responses:
        "200":
          description: Список endpoint'ов сервиса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointListResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/endpoints/{endpointId}:
    get:
      tags:
        - Endpoints
      summary: Получить endpoint
      description: Возвращает данные конкретного endpoint'а.
      operationId: getEndpointById
      parameters:
        - $ref: "#/components/parameters/EndpointId"
      responses:
        "200":
          description: Данные endpoint'а
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointResponse"
        "404":
          description: Endpoint не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Endpoints
      summary: Обновить endpoint
      description: |
        Обновляет параметры endpoint'а: URL, timeout, интервал проверки или признак активности.
      operationId: updateEndpoint
      parameters:
        - $ref: "#/components/parameters/EndpointId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndpointUpdateRequest"
      responses:
        "200":
          description: Endpoint обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointResponse"
        "404":
          description: Endpoint не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/services/{serviceId}/checks:
    get:
      tags:
        - Checks
      summary: Получить историю проверок по сервису
      description: Возвращает сырые результаты проверок endpoint'ов сервиса за период.
      operationId: getServiceCheckHistory
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/CheckStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: История проверок по сервису
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckHistoryResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/endpoints/{endpointId}/checks:
    get:
      tags:
        - Checks
      summary: Получить историю проверок по endpoint'у
      description: Возвращает сырые результаты проверок конкретного endpoint'а за период.
      operationId: getEndpointCheckHistory
      parameters:
        - $ref: "#/components/parameters/EndpointId"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/CheckStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: История проверок по endpoint'у
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckHistoryResponse"
        "404":
          description: Endpoint не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/services/{serviceId}/stats:
    get:
      tags:
        - Statistics
      summary: Получить uptime и SLA по сервису
      description: |
        Рассчитывает uptime и факт нарушения SLA за выбранный период.
        В MVP основной период — month.
      operationId: getServiceStats
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - name: period
          in: query
          required: false
          description: Период расчёта статистики
          schema:
            $ref: "#/components/schemas/StatsPeriod"
            default: month
      responses:
        "200":
          description: Статистика uptime/SLA по сервису
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceStatsResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/services/{serviceId}/summary:
    get:
      tags:
        - Statistics
      summary: Получить сводку по сервису
      description: |
        Возвращает краткую сводку по сервису:
        текущий статус, последний результат проверки, uptime, SLA и последнюю ошибку.
      operationId: getServiceSummary
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: Сводка по сервису
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceSummaryResponse"
        "404":
          description: Сервис не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/services/summary:
    get:
      tags:
        - Statistics
      summary: Получить общую сводку по всем сервисам
      description: Возвращает краткую сводку по всем активным сервисам.
      operationId: getServicesSummary
      responses:
        "200":
          description: Общая сводка по сервисам
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicesSummaryResponse"

components:
  parameters:
    ServiceId:
      name: serviceId
      in: path
      required: true
      description: Идентификатор сервиса
      schema:
        type: string
        format: uuid

    EndpointId:
      name: endpointId
      in: path
      required: true
      description: Идентификатор endpoint'а
      schema:
        type: string
        format: uuid

    DateFrom:
      name: dateFrom
      in: query
      required: false
      description: Начало периода фильтрации
      schema:
        type: string
        format: date-time
        example: "2026-05-01T00:00:00Z"

    DateTo:
      name: dateTo
      in: query
      required: false
      description: Конец периода фильтрации
      schema:
        type: string
        format: date-time
        example: "2026-05-31T23:59:59Z"

    CheckStatus:
      name: status
      in: query
      required: false
      description: Фильтр по статусу проверки
      schema:
        $ref: "#/components/schemas/CheckStatus"

    Limit:
      name: limit
      in: query
      required: false
      description: Количество записей
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100

    Offset:
      name: offset
      in: query
      required: false
      description: Смещение для пагинации
      schema:
        type: integer
        minimum: 0
        default: 0

  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    ServiceCreateRequest:
      type: object
      required:
        - name
        - responsibles
      properties:
        name:
          type: string
          maxLength: 255
          example: orders-service
        description:
          type: string
          nullable: true
          example: Сервис обработки заказов
        slaTargetPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          default: 99.9
          example: 99.9
        responsibles:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ResponsibleCreate"

    ServiceUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: orders-service
        description:
          type: string
          nullable: true
          example: Сервис обработки заказов
        slaTargetPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          example: 99.9
        isActive:
          type: boolean
          example: true

    ServiceResponse:
      type: object
      required:
        - id
        - name
        - slaTargetPercent
        - isActive
        - responsibles
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: orders-service
        description:
          type: string
          nullable: true
          example: Сервис обработки заказов
        slaTargetPercent:
          type: number
          format: double
          example: 99.9
        isActive:
          type: boolean
          example: true
        responsibles:
          type: array
          items:
            $ref: "#/components/schemas/ResponsibleResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ServiceListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceResponse"

    ResponsibleCreate:
      type: object
      required:
        - fullName
        - email
      properties:
        fullName:
          type: string
          maxLength: 255
          example: Иван Петров
        email:
          type: string
          format: email
          example: ivan.petrov@example.com

    ResponsibleResponse:
      type: object
      required:
        - id
        - fullName
        - email
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
          example: Иван Петров
        email:
          type: string
          format: email
          example: ivan.petrov@example.com

    EndpointCreateRequest:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
          maxLength: 255
          example: Orders health
        url:
          type: string
          format: uri
          example: http://test-service:8081/health/up
        method:
          $ref: "#/components/schemas/HttpMethod"
        timeoutMs:
          type: integer
          minimum: 1
          default: 2000
          example: 2000
        checkIntervalSec:
          type: integer
          minimum: 1
          default: 10
          example: 10
        isActive:
          type: boolean
          default: true
          example: true

    EndpointUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: Orders health
        url:
          type: string
          format: uri
          example: http://test-service:8081/health/up
        method:
          $ref: "#/components/schemas/HttpMethod"
        timeoutMs:
          type: integer
          minimum: 1
          example: 2000
        checkIntervalSec:
          type: integer
          minimum: 1
          example: 10
        isActive:
          type: boolean
          example: true

    EndpointResponse:
      type: object
      required:
        - id
        - serviceId
        - name
        - url
        - method
        - timeoutMs
        - checkIntervalSec
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        name:
          type: string
          example: Orders health
        url:
          type: string
          format: uri
          example: http://test-service:8081/health/up
        method:
          $ref: "#/components/schemas/HttpMethod"
        timeoutMs:
          type: integer
          example: 2000
        checkIntervalSec:
          type: integer
          example: 10
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    EndpointListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EndpointResponse"

    CheckResultResponse:
      type: object
      required:
        - id
        - serviceId
        - endpointId
        - checkedAt
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        endpointId:
          type: string
          format: uuid
        checkedAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/CheckStatus"
        httpStatusCode:
          type: integer
          nullable: true
          example: 200
        latencyMs:
          type: integer
          nullable: true
          example: 120
        errorType:
          allOf:
            - $ref: "#/components/schemas/ErrorType"
          nullable: true
        errorMessage:
          type: string
          nullable: true
          example: Request timeout after 2000ms
        createdAt:
          type: string
          format: date-time

    CheckHistoryResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckResultResponse"
        total:
          type: integer
          example: 150
        limit:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0

    ServiceStatsResponse:
      type: object
      required:
        - serviceId
        - serviceName
        - period
        - slaTargetPercent
        - uptimePercent
        - slaBreached
        - totalChecks
        - upChecks
        - downChecks
      properties:
        serviceId:
          type: string
          format: uuid
        serviceName:
          type: string
          example: orders-service
        period:
          $ref: "#/components/schemas/StatsPeriod"
        slaTargetPercent:
          type: number
          format: double
          example: 99.9
        uptimePercent:
          type: number
          format: double
          example: 99.75
        slaBreached:
          type: boolean
          example: true
        totalChecks:
          type: integer
          example: 10000
        upChecks:
          type: integer
          example: 9975
        downChecks:
          type: integer
          example: 25

    ServiceSummaryResponse:
      type: object
      required:
        - serviceId
        - serviceName
        - currentStatus
        - slaTargetPercent
        - uptimePercent
        - slaBreached
        - totalChecks
        - upChecks
        - downChecks
      properties:
        serviceId:
          type: string
          format: uuid
        serviceName:
          type: string
          example: orders-service
        currentStatus:
          $ref: "#/components/schemas/CheckStatus"
        slaTargetPercent:
          type: number
          format: double
          example: 99.9
        uptimePercent:
          type: number
          format: double
          example: 99.75
        slaBreached:
          type: boolean
          example: true
        totalChecks:
          type: integer
          example: 10000
        upChecks:
          type: integer
          example: 9975
        downChecks:
          type: integer
          example: 25
        lastCheckAt:
          type: string
          format: date-time
          nullable: true
        lastError:
          type: string
          nullable: true
          example: HTTP_5XX

    ServicesSummaryResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceSummaryResponse"

    HttpMethod:
      type: string
      description: HTTP-метод проверки endpoint'а. В MVP используется GET.
      enum:
        - GET
      default: GET
      example: GET

    CheckStatus:
      type: string
      description: Статус результата проверки
      enum:
        - UP
        - DOWN
      example: UP

    ErrorType:
      type: string
      description: Тип ошибки при проверке endpoint'а
      enum:
        - HTTP_5XX
        - TIMEOUT
        - CONNECTION_ERROR
        - UNKNOWN_ERROR
      example: HTTP_5XX

    StatsPeriod:
      type: string
      description: Период расчёта статистики
      enum:
        - month
      example: month

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Некорректные входные данные
        details:
          type: object
          additionalProperties: true
          nullable: true